Drop compilers from pure Python packages
Based on discussion and inspection of xgboost, the Python packages
appear to be pure packages. Meaning they do not build C/C++ bindings to
xgboost library. As such they should not need compilers as part of their
build environments. So this attempts to drop the compilers from the pure
parts of the package.

It is worth noting the same is not true of the R packages, which use C++
as part of their bindings. So make sure to leave the compilers for the R
packages.