Add  `github_actions/free_disk_space`
While not officially in `conda-smithy` yet, there is no issue having
this config option here now. Plus it likely will have this name when the
feature lands in a release. So go ahead and add it to ensure we benefit
from this option whenever it is available. Not to mention we can use the
dev version to get this behavior sooner.