Drop CUDA compiler from R package
It appears the R build doesn't use or need the CUDA compiler. Since all
CUDA components were built in the build step and packaged in
`libxgboost`, there is no need for the `r-xgboost` package to rely on
this dependency. So drop it. This aligns with how `py-xgboost` is built
as well.