Slightly relax `libxgboost` pin
Pinning to exact package hash is too stringent as there are now CPU &
CUDA builds of `libxgboost` that are used interchangeably by downstream
XGBoost packages. So drop the package hash constraint, but match the
rest of the version and build string as much as possible. This should
fix some install issues we have seen due to the exact constraint that
was here before.