Template `build/number` and place in header
Simplify bumping the `build/number` by placing it in a Jinja variable at
the top of the recipe. Even though it only is set once, it is simpler
having this at the top next to the version, which may need to be changed
at the same time, as opposed to digging into the recipe to find it.