Add `librmm` to `conda_build_config.yaml`
It seems the `librmm` key is not recognized as something that can be
hashed, which causes `force_use_keys` to fail to find `librmm`. Try
moving `librmm` to `conda_build_config.yaml` instead. This should also
include it in the hash, but is a bit more significant of a recipe
revamp.