Configure `workflow_run` event for upload stub
As GHA will only run the `workflow_run` event for workflows on the
default branch, go ahead and add the event to this stub workflow. That
way it can merged into the default branch without causing issues. We can
then spell out this workflow in a new PR to try and get the behavior we
want.