Fix linter error with `noarch: python`
The linter does not like having `noarch: python` at the top-level with
selectors. Rewriting the recipe without selectors is a substantial lift
and not worth it given the move towards rattler-build.

However if we don't have `noarch` or a `skip` at the top-level, a full
matrix of Python versions is built out for `xgboost` (despite these
packages being redundant).

To clean this up, we skip any Python version except the minimum one that
XGBoost supports. This ensure we build for exactly one Python. To get
the syntax right we need to use Jinja.

Unfortunately that means we need to consolidate our other `skip`s into
this one, which includes the Windows CUDA 11.8 skip. So we do so and try
to make the syntax as pleasing as we can within the constraints.