Fix bot autoupdating parsing error
The autoupdate bot runs into a parsing error with the multiline Jinja
expression used in the top-level `skip`. This places uses a multiline
YAML string to hold the Jinja logic in a value associated with the
`skip` key. This should make sure the bot can parse this logic correctly
without getting hung up on the Jinja contained in it.