Skip if `python_min is undefined`, ensuring it is
Even though `python_min` is used in the recipe, conda-smithy doesn't
seem to always add it in variants. To fix this, skip the case where
`python_min is undefined`. As a result, conda-smithy sees `python_min`
is a used variable and makes sure to add it to all variants. Once added
the skip is effectively moot.