Pin to CCCL 2.7
The `linux_ppc64le` builds are running into issues with CCCL 2.8. The
last successful builds of XGBoost used CCCL 2.7. So just pin to the last
working CCCL build.

Also no need to keep the exclusions for older versions of CCCL. Newer
versions of CCCL are out and would be picked up anyways. So pinning to a
newer CCCL should sufficiently address the same need.

Include a comment with a link back to the build issues seen with CCCL
2.8.