Drop redundant CPU & GPU mutex packages
The same info is already encoded in `build/strings`, which can be easily
used to install the desired variant. The additional mutex and
alternative names of packages merely adds more complexity to building
and maintaining the feedstock. Not to mention provides multiple
duplicative methods for users to install. In the interest of simplifying
the user experience, clean this up so there is one preferred way to
install CPU & GPU packages.