Rely on CMake configuration values for install
As we already configured and built a release build, there should be no
need to specify that at install time. Also the host environment path was
already specified in the build. That same is used for all the split
package installs. So we can just rely on that configuration value when
installing with CMake.