Simplify build string pinning
Previously we would use `libxgboost =*=<build string>`. However it is
preferred to use `libxgboost * <build string>` without the `=`s. This
required a newer version of `conda-build`, which we couldn't use due to
other breaking issues.  Now that we are using a newer `conda-build`
version again.  Try fixing the build string pinning accordingly.